+++ carp_mobile_sensing/lib/domain/transformers.dart
@@ -7,29 +7,25 @@
 part of domain;
 
 /// Signature of a data transformer.
-typedef DataPointTransformer = DataPoint Function(DataPoint);
-
-/// Signature of a data stream transformer.
-typedef DataPointStreamTransformer = Stream<DataPoint> Function(
-    Stream<DataPoint>);
+typedef DatumTransformer = Datum Function(Datum);
 
 /// A no-operation transformer.
-DataPoint noop(DataPoint data) => data;
+Datum noop(Datum datum) => datum;
 
-/// A registry of [DataPointTransformerSchema]s which hold a set of
-/// [DataPointTransformer]s.
-class DataPointTransformerSchemaRegistry {
-  static final DataPointTransformerSchemaRegistry _instance =
-      DataPointTransformerSchemaRegistry._();
+/// A registry of [DatumTransformerSchema]s which hold a set of
+/// [DatumTransformer]s.
+class TransformerSchemaRegistry {
+  static final TransformerSchemaRegistry _instance =
+      TransformerSchemaRegistry._();
 
   /// The map between the namespace of a transformer schema and the schema.
-  Map<String, DataPointTransformerSchema> get schemas => _schemas;
-  final Map<String, DataPointTransformerSchema> _schemas = {};
+  Map<String, DatumTransformerSchema> get schemas => _schemas;
+  final Map<String, DatumTransformerSchema> _schemas = {};
 
-  /// Get the singleton instance of the [DataPointTransformerSchemaRegistry].
-  factory DataPointTransformerSchemaRegistry() => _instance;
+  /// Get the singleton instance of the [TransformerSchemaRegistry].
+  factory TransformerSchemaRegistry() => _instance;
 
-  DataPointTransformerSchemaRegistry._() {
+  TransformerSchemaRegistry._() {
     // register 3 default transformer schemas:
     // 1. a no-operation CARP schema
     // 2. a default OMH schema
@@ -40,64 +36,64 @@ class DataPointTransformerSchemaRegistry {
   }
 
   /// Register a transformer schema.
-  void register(DataPointTransformerSchema schema) {
+  void register(DatumTransformerSchema schema) {
     _schemas[schema.namespace] = schema;
     schema.onRegister();
   }
 
   /// Lookup a transformer schema based on its namespace.
-  DataPointTransformerSchema lookup(String namespace) => _schemas[namespace];
+  DatumTransformerSchema lookup(String namespace) => _schemas[namespace];
 }
 
 /// An interface for Datum that is created from a transformer.
 abstract class TransformedDatum {
-  static DataPointTransformer get transformer => null;
+  static DatumTransformer get transformer => null;
 }
 
 /// An abstract class defining a transformer schema, which hold a set of
-/// [DataPointTransformer]s, which that can map from the native CARP namespace
+/// [DatumTransformer]s, which that can map from the native CARP namespace
 /// to another namespace.
-/// A [DataPointTransformerSchema] must be implemented for each supported namespace.
-abstract class DataPointTransformerSchema {
+/// A [DatumTransformerSchema] must be implemented for each supported namespace.
+abstract class DatumTransformerSchema {
   /// The type of namespace that this package can transform to (see e.g.
   /// [NameSpace] for pre-defined namespaces).
   String get namespace;
 
-  final Map<String, DataPointTransformer> _transformers = {};
+  final Map<String, DatumTransformer> _transformers = {};
 
   /// A map of transformers in this schema, indexed by the data type they
   /// can transform.
-  Map<String, DataPointTransformer> get transformers => _transformers;
+  Map<String, DatumTransformer> get transformers => _transformers;
 
   /// Callback method when this schema is being registered.
   void onRegister();
 
   /// Add a transformer to this schema based on its type mapped to its
   /// [String].
-  void add(String type, DataPointTransformer transformer) =>
+  void add(String type, DatumTransformer transformer) =>
       transformers[type] = transformer;
 
-  /// Transform the [data] according to the transformer for its data type.
-  DataPoint transform(DataPoint data) {
-    Function transformer = transformers[data.carpHeader.dataFormat];
-    return (transformer != null) ? transformer(data) : data;
+  /// Transform the [datum] according to the transformer for its data type.
+  Datum transform(Datum datum) {
+    Function transformer = transformers[datum.format.toString()];
+    return (transformer != null) ? transformer(datum) : datum;
   }
 }
 
-/// A default [DataPointTransformerSchema] for CARP no-operation transformers
-class CARPTransformerSchema extends DataPointTransformerSchema {
+/// A default [DatumTransformerSchema] for CARP no-operation transformers
+class CARPTransformerSchema extends DatumTransformerSchema {
   String get namespace => NameSpace.CARP;
   void onRegister() {}
 }
 
-/// A default [DataPointTransformerSchema] for Open mHealth (OMH) transformers
-class OMHTransformerSchema extends DataPointTransformerSchema {
+/// A default [DatumTransformerSchema] for Open mHealth (OMH) transformers
+class OMHTransformerSchema extends DatumTransformerSchema {
   String get namespace => NameSpace.OMH;
   void onRegister() {}
 }
 
-/// A default [DataPointTransformerSchema] for privacy transformers
-class PrivacySchema extends DataPointTransformerSchema {
+/// A default [DatumTransformerSchema] for privacy transformers
+class PrivacySchema extends DatumTransformerSchema {
   static const String DEFAULT = 'default-privacy-schema';
 
   String get namespace => DEFAULT;