+++ packages/carp_context_package/lib/src/activity/activity_probe.dart
@@ -10,17 +10,12 @@ part of context;
 /// Collects activity information from the underlying OS's activity recognition
 /// API. It generates an [ActivityDatum] every time an activity is detected.
 class ActivityProbe extends StreamProbe {
-  Stream<DataPoint> _stream;
+  Stream<Datum> _stream;
   // Since this probe runs alongside location, which runs a foreground service
   // this probe does not need to run one.
-  Stream<Datum> get stream {
-    if (_stream == null) {
-      _stream = ActivityRecognition.instance
-          .startStream(runForegroundService: false)
-          .where((event) => event.type != ActivityType.UNKNOWN)
-          .map((activity) => ActivityDatum.fromActivity(activity))
-          .asBroadcastStream();
-    }
-    return _stream;
-  }
+  Stream<Datum> get stream => _stream ??= ActivityRecognition.instance
+      .startStream(runForegroundService: false)
+      .where((event) => event.type != ActivityType.UNKNOWN)
+      .map((activity) => ActivityDatum.fromActivity(activity))
+      .asBroadcastStream();
 }