+++ carp_mobile_sensing/CHANGELOG.md
@@ -7,7 +7,7 @@
     * `DataFormat` is now called `DataType`
     * `ManualTrigger` is now called `PassiveTrigger`
     * `ScheduledTrigger` is now called `DateTimeTrigger`
-    * `` is now called ``
+    * `TransformerSchemaRegistry` is now called `DataPointTransformerSchemaRegistry`
     * `` is now called ``
 * A `StudyProtocol` now uses the concept of a `TriggeredTask` which is slightly different to the previous `Study` model. See the example in the README text.
 * All json serialization now used `camelCase` instead of `snake_case` to be consistent with `carp_core`.